@@ -23,10 +23,12 @@
     "@angular/platform-browser-dynamic": "^4.0.0",
     "@angular/platform-server": "^4.0.0",
     "@angular/router": "^4.0.0",
+    "@ngrx/effects": "^4.1.1",
+    "@ngrx/store": "^4.1.1",
     "bootstrap": "^3.3.7",
     "core-js": "^2.4.1",
     "firebase": "^3.7.0",
-    "rxjs": "^5.2.0",
+    "rxjs": "^5.4.2",
     "zone.js": "^0.8.5"
   },
   "devDependencies": {
@@ -46,6 +48,6 @@
     "protractor": "~5.1.0",
     "ts-node": "~2.0.0",
     "tslint": "~4.4.2",
-    "typescript": "~2.1.0"
+    "typescript": "~2.4.2"
   }
 }
