@@ -1,29 +1,45 @@
 import { Router } from '@angular/router';
 import * as firebase from 'firebase';
 import { Injectable } from '@angular/core';
+import { Store } from '@ngrx/store';
+
+import * as fromApp from '../store/app.reducers';
+import * as AuthActions from './store/auth.actions';
+import {SIGNUP} from './store/auth.actions';
 
 @Injectable()
 export class AuthService {
-  token: string;
-
-  constructor(private router: Router) {}
+  constructor(private router: Router, private store: Store<fromApp.AppState>) {}
 
   signupUser(email: string, password: string) {
     firebase.auth().createUserWithEmailAndPassword(email, password)
+      .then(
+        user => {
+        this.store.dispatch(new AuthActions.Signup());
+          firebase.auth().currentUser.getToken()
+            .then(
+              (token: string) => {
+                this.store.dispatch(new AuthActions.SetToken(token));
+              }
+            )
+      })
       .catch(
         error => console.log(error)
-      )
+      );
   }
 
   signinUser(email: string, password: string) {
     firebase.auth().signInWithEmailAndPassword(email, password)
       .then(
         response => {
+          this.store.dispatch(new AuthActions.Signin());
           this.router.navigate(['/']);
           firebase.auth().currentUser.getToken()
             .then(
-              (token: string) => this.token = token
-            )
+              (token: string) => {
+                this.store.dispatch(new AuthActions.SetToken(token));
+              }
+            );
         }
       )
       .catch(
@@ -33,18 +49,6 @@ export class AuthService {
 
   logout() {
     firebase.auth().signOut();
-    this.token = null;
-  }
-
-  getToken() {
-    firebase.auth().currentUser.getToken()
-      .then(
-        (token: string) => this.token = token
-      );
-    return this.token;
-  }
-
-  isAuthenticated() {
-    return this.token != null;
+    this.store.dispatch(new AuthActions.Logout());
   }
 }
