@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Http, Response } from '@angular/http';
+import { HttpClient, HttpParams, HttpRequest } from '@angular/common/http';
 import 'rxjs/Rx';
 
 import { RecipeService } from '../recipes/recipe.service';
@@ -8,25 +8,30 @@ import { AuthService } from '../auth/auth.service';
 
 @Injectable()
 export class DataStorageService {
-  constructor(private http: Http,
-              private recipeService: RecipeService,
-              private authService: AuthService) {
+  URL = 'https://recipesbook-845f3.firebaseio.com/recipes.json';
+  constructor(private http: HttpClient,
+              private recipeService: RecipeService) {
   }
 
   storeRecipes() {
-    const token = this.authService.getToken();
-
-    return this.http.put('https://ng-recipe-book.firebaseio.com/recipes.json?auth=' + token, this.recipeService.getRecipes());
+    // return this.http.put('https://ng-recipe-book.firebaseio.com/recipes.json',
+    //   this.recipeService.getRecipes(), {
+    //     observe: 'body',
+    //     params: new HttpParams().set('auth', token)
+    //   });
+    const req = new HttpRequest('PUT', this.URL,
+      this.recipeService.getRecipes(), {reportProgress: true});
+    return this.http.request(req);
   }
 
   getRecipes() {
-    const token = this.authService.getToken();
-
-    this.http.get('https://ng-recipe-book.firebaseio.com/recipes.json?auth=' + token)
+    this.http.get<Recipe[]>(this.URL, {
+      observe: 'body', // default
+      responseType: 'json', // default
+    })
       .map(
-        (response: Response) => {
-          const recipes: Recipe[] = response.json();
-          for (let recipe of recipes) {
+        (recipes) => {
+          for (const recipe of recipes) {
             if (!recipe['ingredients']) {
               recipe['ingredients'] = [];
             }
